### 2021-07-10 #3469 @Kevin-Mattheus-Moerman
@patrickfuchs thank you for your submission to JOSS. Can you please check those invalid DOI's :point_up:
Note that you can call `@whedon generate pdf` to update the paper and `@whedon check references` to update the reference check.
