### 2021-07-10 #3469 @richardjgowers
Yeah I can take this.

On Sat, Jul 10, 2021 at 07:58, Kevin Mattheus Moerman <
***@***.***> wrote:

> @patrickfuchs <https://github.com/patrickfuchs> I have added the
> waitlisted label to this submission as all editors in this domain are
> currently preoccupied with other submissions. We will assign an editor once
> one becomes available.
>
> @richardjgowers <https://github.com/richardjgowers> perhaps you could
> handle this submission once you are available?
>
