+++ paper/paper.bib
@@ -61,7 +61,7 @@ number = {1},
 pages = {43-56},
 year = {1995},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/0010-4655(95)00042-E},
+doi = {10.1016/0010-4655(95)00042-E},
 url = {https://www.sciencedirect.com/science/article/pii/001046559500042E},
 author = {H.J.C. Berendsen and D. {van der Spoel} and R. {van Drunen}},
 keywords = {Molecular dynamics, Parallel computing}
@@ -74,7 +74,7 @@ volume = {1-2},
 pages = {19-25},
 year = {2015},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2015.06.001},
+doi = {10.1016/j.softx.2015.06.001},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711015000059},
 author = {Mark James Abraham and Teemu Murtola and Roland Schulz and Szilárd Páll and Jeremy C. Smith and Berk Hess and Erik Lindahl},
 keywords = {Molecular dynamics, GPU, SIMD, Free energy}
@@ -102,7 +102,7 @@ number = {5},
 pages = {1727-1739},
 year = {1995},
 issn = {0006-3495},
-doi = {https://doi.org/10.1016/S0006-3495(95)80350-4},
+doi = {10.1016/S0006-3495(95)80350-4},
 url = {https://www.sciencedirect.com/science/article/pii/S0006349595803504},
 author = {J.P. Douliez and A. Léonard and E.J. Dufourc}
 }
@@ -126,7 +126,7 @@ number = {5},
 pages = {2002-2013},
 year = {1997},
 issn = {0006-3495},
-doi = {https://doi.org/10.1016/S0006-3495(97)78845-3},
+doi = {10.1016/S0006-3495(97)78845-3},
 url = {https://www.sciencedirect.com/science/article/pii/S0006349597788453},
 author = {O. Berger and O. Edholm and F. Jähnig}
 }
@@ -177,7 +177,7 @@ volume = {32},
 number = {10},
 pages = {2319-2327},
 keywords = {molecular dynamics simulations, analysis, proteins, object-oriented design, software, membrane systems, Python programming language},
-doi = {https://doi.org/10.1002/jcc.21787},
+doi = {10.1002/jcc.21787},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21787},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21787},
 year = {2011}
@@ -264,7 +264,7 @@ number = {3},
 pages = {235-270},
 year = {1997},
 issn = {0304-4157},
-doi = {https://doi.org/10.1016/S0304-4157(97)00008-7},
+doi = {10.1016/S0304-4157(97)00008-7},
 url = {https://www.sciencedirect.com/science/article/pii/S0304415797000087},
 author = {D.P Tieleman and S.J Marrink and H.J.C Berendsen}
 }
@@ -277,7 +277,7 @@ number = {3},
 pages = {217-223},
 year = {2000},
 issn = {1359-0294},
-doi = {https://doi.org/10.1016/S1359-0294(00)00058-3},
+doi = {10.1016/S1359-0294(00)00058-3},
 url = {https://www.sciencedirect.com/science/article/pii/S1359029400000583},
 author = {Scott E. Feller},
 keywords = {Molecular dynamics, Simulation, Force-field, Algorithm, Interactions}
@@ -312,5 +312,5 @@ url="https://doi.org/10.1007/978-1-59745-519-0_7"
[…]
