### 2021-07-22 #3521 @whedon
[REVIEW]: buildH: Build hydrogen atoms from united-atom molecular dynamics of lipids and calculate the order parameters
**Submitting author:** @patrickfuchs (<a href="http://orcid.org/0000-0001-7117-994X">Patrick Fuchs</a>)
**Repository:** <a href="https://github.com/patrickfuchs/buildH" target ="_blank">https://github.com/patrickfuchs/buildH</a>
**Version:** v1.6

- [x] #3667
