### 2021-08-09 d5a35c4
Fixing box dimensions and trajectory output with Hs
