### 2021-08-09 patrickfuchs/buildH#145 @patrickfuchs
Fixing box dimensions and trajectory output with Hs
In PR #142, this line proposed a way to write box dimensions in the xtc:
https://github.com/patrickfuchs/buildH/blob/1a2c740136910b3293dde395cb37e76ecb566454/buildh/core.py#L649

Unfortunately, for some reason, it *rewinds the trajectory* to the first frame so that the newly built xtc with Hs gets the same (first) frame written at each iteration of the loop!!!
In commit d5a35c4, I fixed that with

```
universe_wH.trajectory.ts.dimensions = ts.dimensions
```

which avoids this rewind. I also added box dimensions in the generated pdb with Hs + a remark in case no unitcell info was present in the input pdb.

The commit above should fix this issue, but I leave it open until you (@HubLot & @pierrepo) are back from vacation, so you can have a look 😄.
