### 2021-08-09 patrickfuchs/buildH#145 @patrickfuchs
I also added some tests (commit ac6303d) to check the integrity of the newly generated pdb and xtc with Hs.
