### 2021-08-10 7441be9
Fix environment.yml for pip
At the end of July, the format for installing requirements from a file has changed
Source: https://stackoverflow.com/questions/68571543/using-a-pip-requirements-file-in-a-conda-yml-file-throws-attributeerror-fileno
