### 2021-08-10 4dbe4a8
Fix menu & add credit to Durand's blog post
