### 2021-08-28 patrickfuchs/buildH#145 @richardjgowers
@patrickfuchs just to clarify this behaviour.  MDAnalysis only holds a single frame in memory at any time (with memoryreader being the exception) `u.trajectory[n]` is the syntax for loading frame `n` of a trajectory.  It's worth noting that changes to `.dimensions` / `.positions` won't be persistent if you move between frames.  Let me know if you have any more questions on this behaviour.
