### 2021-08-28 #3521 @richardjgowers
Hi @ojeda-e @blakeaw @lilyminium looks like we've made some progress on the review.  For remaining points you'd like to raise, if you can either raise issues on the repo (which I see we've already done) or directly ask the author in this thread (@patrickfuchs)  then we can move this review along.
