### 2021-08-28 patrickfuchs/buildH#148 @lilyminium
Which OSs are supported?
I see that the tests run on Ubuntu, so I could reasonably expect buildH to work on Linux and Mac OSs (although adding even 1 test case for Mac would help with looking for bugs). Just curious, is Windows supported at all, and if so, will you add CI? I see that the conda package is `noarch`.
