### 2021-08-28 patrickfuchs/buildH#149 @lilyminium
Minor wording or grammar nitpicks with paper
Thank you for your paper. I have listed here a number of minor nitpicks with the wording or grammar, followed by suggestions. These are all just suggestions, but I have bolded any words or grammar that I really believe should be changed for clarity. Please let me know if I should give more information anywhere.

> In AA force fields (such as CHARMM36 [@Klauda2010]), all the atoms are considered whereas in UA force fields (such as Berger [@Berger1997]) the aliphatic hydrogen atoms (Hs) are merged **to** their parent carbon into a larger particle representing a CH, CH2 or CH3 (e.g. a methyl group is represented by a single CH3 particle).

Suggestion: "the aliphatic hydrogen atoms (Hs) are merged **with** their parent carbon into a single, larger particle (e.g. a methyl group is represented by a CH3 particle)."

> In simulations of phospholipids, the use of UA representations **allows to divide by almost 3** the number of **atoms** to simulate because these molecules contain many aliphatic Hs.

Suggestion: "Phospholipids contain many aliphatic Hs, so the use of UA representations in simulation allows us to **reduce** the number of **particles** that are simulated to almost a third of the true number of **atoms**."

> In this type of experiment, aliphatic **Hs atoms** are replaced by deuterons.

Suggestion: "H atoms" *or* "Hs" alone

> It describes the amount of possible orientations visited by the C-H bond, varies between 1 and -0.5 and is unitless.

A comma after "varies between 1 and -0.5" would help in separating it from the "unitless". Also, suggestion: "ranges between -0.5 and 1, inclusive" ?

> Values close to 0 mean high mobility, when it goes away from 0 (towards negative or positive values), it means the C-H bond gets less mobile.

Suggestion: "Values close to 0 indicate high mobility, while values with higher magnitudes (either positive or negative) indicate that the C-H bond is less mobile."

>  ii) reconstructing Hs coordinates and calculat**e** $S_{CH}$ as in AA simulations
Suggestion: "calculat**ing**"

I would also find it easier to just spell out "reconstructing hydrogen coordinates"

> The trend in the last years **was more to use strategy ii)**, such as in the NMRlipids project

Suggestion: "The trend in the recent years has been towards strategy ii)"

> **making** fundamental research on the structure and dynamics of lipid membranes.

"conducting"?

> integrate in analyses pipelines or Jupyter notebooks

analyses -> analysis ?
