### 2021-08-28 #3521 @lilyminium
Thank you for your patience @patrickfuchs and @richardjgowers, sorry that the review has taken so long. I'm only waiting for a clarification as to the supported OSs (https://github.com/patrickfuchs/buildH/issues/148) and just posted some minor nitpicks with the paper (https://github.com/patrickfuchs/buildH/issues/149), but either way I am happy to recommend this for publication. `buildH` is well-documented and easy-to-use, especially with "standard" lipids such as the Berger set. It is clearly targeted towards GROMACS users; tutorials are written for TRR and XTC formats, and the output is only available in XTC. (I note that with the use of MDAnalysis, it would be easy to expand the supported formats beyond the current recommended options of PDB/GRO + XTC/TRR/DCD.) Many of the flags will be familiar to gromacs users (-c, -o, -b/-e). With the disappearance of `g_protonate`, I believe that this will be a welcome tool for analysis of lipid simulations.
