### 2021-08-28 patrickfuchs/buildH#150 @ojeda-e
Improvements on API documentation
I have a couple of suggestions to improve the readability of the docs.

1. API [geometry](https://buildh.readthedocs.io/en/latest/api/geometry.html).

- The [order parameter equation](https://github.com/patrickfuchs/buildH/blob/4dbe4a81348c2be10e0e76ccbbc6b0722e1d9789/buildh/geometry.py#L180) is displayed as text. Suggestion: Using the `.. math::`  sphinx markup for equations will make it more readable.
- The diagram in docstrings, from lines [184-188](https://github.com/patrickfuchs/buildH/blob/4dbe4a81348c2be10e0e76ccbbc6b0722e1d9789/buildh/geometry.py#L184), doesn't display properly. Given the aim of the package, displaying it properly is relevant. Adding backquotes may help to fix it.

2. API [hydrogens](https://buildh.readthedocs.io/en/latest/api/hydrogens.html)
- In docstrings, all the return types are plain text. In general, backquotes are used for inline code.
Example: For lines [29-32 in hydrogens.py](https://github.com/patrickfuchs/buildH/blob/4dbe4a81348c2be10e0e76ccbbc6b0722e1d9789/buildh/hydrogens.py#L32)
Suggested:
```
    Returns
    -------
    numpy 1D-array
        Coordinates of the rebuilt hydrogen: ``([x_H, y_H, z_H])``.
```

3.  API [init_dics](https://buildh.readthedocs.io/en/latest/api/init_dics.html)
- As in API hydrogens, use backquotes (``) or (```) for inline code.
For example,  [lines 54-58](https://github.com/patrickfuchs/buildH/blob/4dbe4a81348c2be10e0e76ccbbc6b0722e1d9789/buildh/init_dics.py#L54) should be inline code.

4. API [lipids](https://buildh.readthedocs.io/en/latest/api/lipids.html) and [utils](https://buildh.readthedocs.io/en/latest/api/utils.html)
- In `buildh.lipids.check_topology`, the type of dict is `dict` but in previous API pages appears as `dictionary`. For consistency, it would be better to have one only.  Same for the types of parameters in `buildh.utils.is_allHs_present` and `buildh.utils.check_slice_options` (`dict` vs `dictionary`).
