### 2021-08-28 patrickfuchs/buildH#151 @ojeda-e
Python versions and dependencies.
The [Requirements](https://github.com/patrickfuchs/buildH#requirements) section in the README says:

> buildH is written in Python 3 and needs the modules numpy, pandas, MDAnalysis and Numba.

- Is there a way to know the minimum versions of the dependencies?
- numpy -> NumPy.

> Python >= 3.6 and <= 3.8 is mandatory for running buildH
- Since the repo doesn't include CI, out of curiosity, how did you identify Python 3.6 as min and 3.8 as max versions?
