### 2021-08-28 #3521 @ojeda-e
Apologies for the delay @patrickfuchs, @richardjgowers
The statement of need is clear and the authors make a valid point on how `buildH` fills a gap in the set of available MD analysis tools given the interesting evolution of GROMACS over the years. Reconstructing missing hydrogen atoms together with their trajectories will be particularly useful to investigate lipids at the molecular level and allows broader comparisons with MD simulations carried out with other force fields. New users can easily familiarize themselves with the tool following the tutorials provided.
I raised some minor points on the API documentation (issue [#150](https://github.com/patrickfuchs/buildH/issues/150)), where I also offered some suggestions intended to improve an already good and readable API. Another point I asked for clarification is how the min and max versions of Python were determined without CI (issue [#151](https://github.com/patrickfuchs/buildH/issues/151)) as well as the minimum requirements of the dependencies.
I support its publication. `buildH` is easy to use and GROMACS users will find value in it.
