### 2021-08-30 patrickfuchs/buildH#151 @pierrepo
Hi @ojeda-e

> Is there a way to know the minimum versions of the dependencies?

Not for all. However, MDAnalysis is the strongest dependency. We need MDAnalysis version >= 1.1.1 due to some annoying warning messages in previous versions. MDAnalysis itself requires [NumPy >= 1.16.0](https://github.com/MDAnalysis/mdanalysis/blob/develop/package/setup.py).

> Since the repo doesn't include CI, out of curiosity, how did you identify Python 3.6 as min and 3.8 as max versions?

We have a CI as GitHub actions, see for instance in the `.github/workflows` directory.

We need Python 3.6 because we are using f-strings and Python <= 3.8 is a requirement of MDAnalysis 1.1.1. However, MDAnalysis 2.0.0 has been released a week ago and is compatible from Python 3.6 to Python 3.9. We might test compatibility with MDanalysis 2.0.0.
