### 2021-08-30 patrickfuchs/buildH#145 @patrickfuchs
Thanks very much @richardjgowers for this clarification, this is very useful. We made the change using the syntax `u.trajectory[n]` just before submission (to avoid some warnings from MDAnalysis), but were not aware of this behavior. I think now we are fine.
