### 2021-08-30 patrickfuchs/buildH#145 @HubLot
well spotted @patrickfuchs!
I don't understand how the previous code was supposed to work now I see the fix...
This bug,  by itself, should trigger a minor version upgrade.
