### 2021-08-30 e67622a
Simplify workflow file and add Python 3.9 for CI
