### 2021-08-30 2df1091
Improve docstrings
