+++ buildh/core.py
@@ -27,7 +27,7 @@ def buildHs_on_1C(atom, H_type, helper1, helper2, helper3=None):
         Central atom on which we want to reconstruct the hydrogen.
     H_type: str
         The type of H to build. It could be 'CH2', 'CH', 'CHdoublebond' or 'CH3'
-        see dic_lipds.py
+        (see `init_dics.py`)
     helper1 : numpy 1D-array
         First neighbor of central atom.
     helper2 : numpy 1D-array
@@ -40,7 +40,7 @@ def buildHs_on_1C(atom, H_type, helper1, helper2, helper3=None):
     tuple of numpy 1D-arrays
         Each element of the tuple is a numpy 1D-array containing 1, 2 or 3
         reconstructed hydrogen(s).
-        !!! IMPORTANT !!! This function *should* return a tuple even if
+        **Important**: This function return a tuple even if
         there's only one H that has been rebuilt.
     """
     if H_type == "CH2":
@@ -67,36 +67,32 @@ def buildHs_on_1C(atom, H_type, helper1, helper2, helper3=None):
 
 
 def build_system_hydrogens(universe_woH, dic_lipid, dic_Cname2Hnames, dic_lipid_indexes):
-    """Build a new system *with* hydrogens.
+    """Build a new system **with** hydrogens.
 
-    The function takes the MDAnalysis universe *without* hydrogens, reconstructs all hydrogens
+    The function takes the MDAnalysis universe **without** hydrogens, reconstructs all hydrogens
     and returns a pandas dataframe. This latter will be used later to build a new
     MDAnalysis universe with H.
 
-    Notes
-    -----
+    Note
+    ----
     There is no simple way to create a new MDAnalysis universe directly.
 
     Parameters
     ----------
     universe_woH : MDAnalysis universe instance
-        This is the universe *without* hydrogen.
+        This is the universe **without** hydrogen.
     dic_lipid : dictionary
-        Comes from dic_lipids.py. Contains carbon names and helper names needed
-        for reconstructing hydrogens.
+        Contains carbon names and helper names needed for reconstructing hydrogens.
     dic_Cname2Hnames : dictionary
-        This dict gives the correspondance Cname -> Hname. It is a dict of
-        tuples. If there is more than 1 H for a given C, they need to be
-        *ordered* like in the PDB. e.g. for CHARMM POPC :
-        {'C13': ('H13A', 'H13B', 'H13C'), ..., 'C33': ('H3X', 'H3Y'),
-          ..., 'C216': ('H16R', 'H16S'), ...}
+        This dict gives the correspondance Cname -> Hname.
+        (see `make_dic_Cname2Hnames()` in `init_dics.py`)
     dic_lipids_with_indexes : dictionary
-        The dictionary made in function make_dic_lipids_with_indexes().
+        The dictionary made in function `make_dic_lipids_with_indexes()`.
 
     Returns
     -------
     pandas dataframe
-        contains the system *with* hydrogens.
+        contains the system **with** hydrogens.
     """
     # The list newrows will be used to store the new molecule *with* H.
     newrows = []
@@ -164,49 +160,44 @@ def build_all_Hs_calc_OP(universe_woH, ts, dic_lipid, dic_Cname2Hnames, universe
                          dic_corresp_numres_index_dic_OP, dic_lipid_indexes):
     """Build all hydrogens and calculates order parameters for one frame.
 
-    This function loop overs *all* atoms of the universe_woH in order to update
+    This function loop overs **all** atoms of the universe_woH in order to update
     the atom coordinates and the new H built into the universe_wH.
 
     The function also calculates the order parameter.
-    The coordinates of the universe *with* H are updated in place.
[…]
