### 2021-08-30 #3521 @blakeaw
I apologize for the delay. Thank you for your patience, @patrickfuchs, @richardjgowers.
I grade the submission as category 1) Accept.

The authors provide a clear statement of need and give a nice overview of the relevant background. The code is quite well documented, and it was easy to install and use. I think buildH will be a useful analysis tool for anyone simulating lipids with united atom force fields.
