### 2021-08-30 5294e91
Amend text as suggested in PR#149
