+++ paper/paper.bib
@@ -313,5 +313,5 @@ url="https://doi.org/10.1007/978-1-59745-519-0_7"
 title="Inverse conformational selection in lipid-protein binding"
 author="Am{\'e}lie Bacle and Pavel Buslaev and Rebeca Garc{\'i}a Fandi{\~n}o and Fernando Favela-Rosales and Tiago M. Ferreira and Patrick F.J. Fuchs and Ivan Gushchin and Matti Javanainen and Anne M. Kiirikki and Jesper J. Madsen and Josef Melcr and Paula Mil{\'a}n Rodr{\'i}guez and Markus S. Miettinen and O. H. Samuli Ollila and Chris G. Papadopoulos and Antonio Pe{\'o}n and Thomas J. Piggot and {\'A}ngel Pi{\~n}eiro and Salla I. Virtanen",
 year="2021",
-doi="10.5281/zenodo.4748646"
+doi="10.1021/jacs.1c05549"
 }
