+++ paper/paper.md
@@ -46,7 +46,7 @@ $$S_{CH} = \frac{1}{2} \left \langle 3cos^2(\theta) -1 \right \rangle$$
 
 where $\theta$ is the angle between the C-H bond and the magnetic field. The symbol $\langle ... \rangle$ means averaging over time and molecules.
 
-This order parameter is useful because it is directly related to the flexibility of the given C-H bond. It describes the amount of possible orientations visited by the C-H bond, ranges between -0.5 and 1 (inclusive), and is unitless. Values close to 0 indicate high mobility, while values with higher magnitudes (either positive or negative) indicate that the C-H bond is less mobile. Although it varies theoretically between 1 and -0.5, its absolute value $\lvert S_{CH} \rvert$ is often reported because its sign is usually difficult to measure experimentally [@Ollila2014]. In MD simulations, $\theta$ is the angle between the C-H bond and the normal to the membrane (usually the $z$ axis). For AA simulations, $S_{CH}$ is trivial to calculate. However, it is more difficult for UA simulations since the aliphatic Hs are not present. There are two main strategies to compute $S_{CH}$ from UA simulations [@Piggot2017]: i) expressing $S_{CH}$ as a function of the coordinates of other atoms [@Douliez1995], ii) reconstructing hydrogen coordinates and calculating $S_{CH}$ as in AA simulations. The trend in the recent years has been towards strategy ii), such as in the NMRlipids project [@Botan2015; @Catte2016; @Antila2019; @Bacle2021]. NMRlipids is an open science project which uses MD simulations and experimental $S_{CH}$ with the goal of improving lipid force fields or conducting fundamental research on the structure and dynamics of lipid membranes.
+This order parameter is useful because it is directly related to the flexibility of the given C-H bond. It describes the amount of possible orientations visited by the C-H bond, ranges between -0.5 and 1 (inclusive), and is unitless. Values close to 0 indicate high mobility, while values with higher magnitudes (either positive or negative) indicate that the C-H bond is less mobile. Although it varies theoretically between -0.5 and 1, its absolute value $\lvert S_{CH} \rvert$ is often reported because its sign is usually difficult to measure experimentally [@Ollila2014]. In MD simulations, $\theta$ is the angle between the C-H bond and the normal to the membrane (usually the $z$ axis). For AA simulations, $S_{CH}$ is trivial to calculate. However, it is more difficult for UA simulations since the aliphatic Hs are not present. There are two main strategies to compute $S_{CH}$ from UA simulations [@Piggot2017]: i) expressing $S_{CH}$ as a function of the coordinates of other atoms [@Douliez1995], ii) reconstructing hydrogen coordinates and calculating $S_{CH}$ as in AA simulations. The trend in the recent years has been towards strategy ii), such as in the NMRlipids project [@Botan2015; @Catte2016; @Antila2019; @Bacle2021]. NMRlipids is an open science project which uses MD simulations and experimental $S_{CH}$ with the goal of improving lipid force fields or conducting fundamental research on the structure and dynamics of lipid membranes.
 
 # Statement of Need
 
