### 2021-08-31 patrickfuchs/buildH#148 @patrickfuchs
> I see that the tests run on Ubuntu, so I could reasonably expect buildH to work on Linux and Mac OSs (although adding even 1 test case for Mac would help with looking for bugs). Just curious, is Windows supported at all, and if so, will you add CI? I see that the conda package is `noarch`.

Thanks for your comment. We didn't plan to check different OS versions. In principle, buildH is pure Python so it should work on any platform provided its dependencies are supported. We will add a note in the doc about this.
