### 2021-08-31 #3521 @patrickfuchs
Many thanks @richardjgowers for editing the manuscript, and to @ojeda-e @blakeaw @lilyminium for their reviewing and useful suggestions. My collaborators and I are glad they overall recommend acceptance. We answered / adressed the issues opened by @lilyminium and @ojeda-e, we are now waiting whether they are satisfied. We are also grateful to all of you for some useful explanations on how MDAnalysis works.
Furthermore, since MDAnalysis 2.0.0 was released during the review process and is now automatically installed with pip, we made a few changes so that buildH works now with this new version.
