### 2021-08-31 #3521 @patrickfuchs
One other question @richardjgowers. We recently corrected an important [bug regarding box dimensions](https://github.com/patrickfuchs/buildH/issues/145), and buildH is now compatible with MDAnalysis 2.0.0 as stated above. So we think it deserves a new release (from 1.5 to 1.6) that would be sticked to the JOSS paper. When do you advice to make this new release, now or later in the review process?
