### 2021-08-31 patrickfuchs/buildH#150 @patrickfuchs
Thanks for the useful suggestions @ojeda-e. @HubLot will implement all of these within a couple of days.
