### 2021-08-31 patrickfuchs/buildH#150 @HubLot
Thanks @ojeda-e for your detailed issue. I create a PR (#154) which should fix all the points raised and some others too I spotted during reviewing the docstrings.
