### 2021-08-31 4aa0f4e
small typos in docstrings
