+++ paper/paper.bib
@@ -311,7 +311,7 @@ url="https://doi.org/10.1007/978-1-59745-519-0_7"
 
 @article{Bacle2021,
 title="Inverse conformational selection in lipid-protein binding"
-author="Amélie Bacle and Pavel Buslaev and Rebeca Garca Fandino and Fernando Favela-Rosales and Tiago M. Ferreira and Patrick F.J. Fuchs and Ivan Gushchin and Matti Javanainen and Anne M. Kiirikki and Jesper J. Madsen and Josef Melcr and Paula Milan Rodriguez and Markus S. Miettinen and O. H. Samuli Ollila and Chris G. Papadopoulos and Antonio Peon and Thomas J. Piggot and Angel Pineiro and Salla I. Virtanen",
+author="Amélie Bacle and Pavel Buslaev and Rebeca Garcia Fandino and Fernando Favela-Rosales and Tiago M. Ferreira and Patrick F.J. Fuchs and Ivan Gushchin and Matti Javanainen and Anne M. Kiirikki and Jesper J. Madsen and Josef Melcr and Paula Milan Rodriguez and Markus S. Miettinen and O. H. Samuli Ollila and Chris G. Papadopoulos and Antonio Peon and Thomas J. Piggot and Angel Pineiro and Salla I. Virtanen",
 year="2021",
 doi="10.1021/jacs.1c05549"
 }
