### 2021-08-31 #3521 @richardjgowers
@patrickfuchs you can cut a release during the review process and we can change the target of this review to 1.6, the changes you are describing shouldn't be too much for the reviewers to work with.
