### 2021-08-31 #3521 @patrickfuchs
OK thanks very much for your answer @richardjgowers. We will thus soon release the 1.6 version.
