### 2021-09-01 d58f1e4
Add section requirements and compatibility
