+++ setup.py
@@ -1,3 +1,3 @@
 import setuptools
 
-setuptools.setup(version="1.5.0")
+setuptools.setup(version="1.6.0")
