+++ devtools/bumpversion.cfg
@@ -1,7 +1,7 @@
 [bumpversion]
 commit = True
 tag = True
-current_version = 1.5.0
+current_version = 1.6.0
 
 [bumpversion:file:buildh/__init__.py]
 
