### 2021-09-01 patrickfuchs/buildH#151 @patrickfuchs
With the new version 1.6.0 just released, buildH is now fully compatible with MDAnalysis 2.0.0 and Python 3.9.
