### 2021-09-01 #3521 @patrickfuchs
> @patrickfuchs you can cut a release during the review process and we can change the target of this review to 1.6, the changes you are describing shouldn't be too much for the reviewers to work with.

@richardjgowers I just released version 1.6, would it possible to make it the new target of this review? Thanks in advance.
