### 2021-09-02 patrickfuchs/buildH#151 @ojeda-e
Thanks for your answers, @pierrepo @patrickfuchs.
I missed your CI, my apologies. I see the tests for Python 3.6, 3.7, 3.8, and 3.9. I also found minimum versions of dependencies in the [documentation](https://buildh.readthedocs.io/en/latest/installation.html), which makes everything more clear.
Thanks for adding these changes.
