### 2021-09-06 #3521 @patrickfuchs
Thank you @lilyminium and @ojeda-e.  Just to add a word regarding CI, I see one box is unchecked, we have automated tests implemented through [github actions](https://github.com/patrickfuchs/buildH/actions), see [here](https://github.com/patrickfuchs/buildH/tree/master/.github/workflows) for the specifications.
@richardjgowers, let me know if we need to do any other thing.
