### 2021-09-12 #3521 @richardjgowers
@patrickfuchs the only thing I can see is that the reference to the latest version needs to be updated to now say v1.6.  After that can you redo the Zenodo deposit and report what the DOI for this is?
