### 2021-09-16 a5332af
Update buildH version + Zenodo link
