### 2021-09-17 #3521 @patrickfuchs
@richardjgowers, we updated the links to the Zenodo and Software Heritage archive, both now point to the current version, i.e. 1.6.0. The new Zenodo link is : https://zenodo.org/record/5356246.
