### 2021-09-20 #3521 @patrickfuchs
On behalf of my colleagues and myself, we would like to thank the editor @richardjgowers and the reviewers @lilyminium, @ojeda-e, @blakeaw. This was a nice review experience!
@arfon, I have a last question. I spotted a couple of typos in the paper, is it possible to amend the paper at this stage? If not, this is not a big deal (these are really minor things : typos in the full name of lipids, one missing journal name in a ref but the DOI is OK).
