### 2021-09-20 #3521 @arfon
Thanks for the positive feedback @patrickfuchs!

> I spotted a couple of typos in the paper, is it possible to amend the paper at this stage? If not, this is not a big deal (these are really minor things : typos in the full name of lipids, one missing journal name in a ref but the DOI is OK).

Certainly. Please make the changes to the paper in your repository and let me know when you're ready for me to update the paper.
