### 2021-09-21 #3521 @patrickfuchs
> Certainly. Please make the changes to the paper in your repository and let me know when you're ready for me to update the paper.

Thanks very much @arfon for your prompt and positive answer! I pushed the changes on my repository so the paper is ready to be updated.
