### 2021-09-27 #3521 @arfon
This should be done now. The paper might take a few hours to update on the JOSS website.
